verify refactor

+++ GameAPI/Program.cs
@@ -1,3 +1,4 @@
+using System.Collections.Specialized;
 using System.Security.Cryptography;
 using System.Text;
 using System.Text.Json;
@@ -126,16 +127,19 @@ app.MapPost("/api/players", async ([FromBody] Player player, PlayerService playe
     return Results.CreatedAtRoute("GetPlayer", new { id = player.TelegramId }, player);
 });
 
-app.MapPut("/api/players/{id:long}", async (long id, Player playerIn, PlayerService playerService, ICacheService cacheService) =>
+app.MapPut("/api/players/{id:long}", async (long id, Player playerIn, PlayerService playerService,
+    ICacheService cacheService) =>
 {
     var player = await playerService.GetAsync(id);
-    
+    if(player is null)
+        return Results.NotFound();
     await playerService.UpdateAsync(id, playerIn);
     await cacheService.RemoveAsync(cacheKey);  // Invalidate cache
-    return Results.NotFound();
+    return Results.Ok();
 });
 
-app.MapPut("/api/players/{telegramId:long}/rating", async (long telegramId, [FromBody] int ratingChange, PlayerService playerService, ICacheService cacheService) =>
+app.MapPut("/api/players/{telegramId:long}/rating", async (long telegramId, [FromBody] int ratingChange,
+    PlayerService playerService, ICacheService cacheService) =>
 {
     var success = await playerService.UpdateRatingAsync(telegramId, ratingChange);
 
@@ -150,7 +154,8 @@ app.MapPut("/api/players/{telegramId:long}/rating", async (long telegramId, [Fro
     }
 });
 
-app.MapDelete("/api/players/{id:long}", async (long id, PlayerService playerService, ICacheService cacheService) =>
+app.MapDelete("/api/players/{id:long}", async (long id, PlayerService playerService,
+    ICacheService cacheService) =>
 {
     var player = await playerService.GetAsync(id);
 
@@ -169,7 +174,8 @@ app.MapGet("/api/leaders", async ([FromServices] ICacheService cacheService) =>
             return Results.NotFound("No leaders found.");
         }
         
-        var playerList = leaders.Select(leader => new Player { TelegramId = leader.Key, Rating = (int)leader.Value }).ToList();
+        var playerList = leaders.Select(leader => new Player { TelegramId = leader.Key,
+            Rating = (int)leader.Value }).ToList();
 
         return Results.Ok(playerList);
     }
@@ -182,65 +188,21 @@ app.MapGet("/api/leaders", async ([FromServices] ICacheService cacheService) =>
 
 app.MapPost("/api/verify", async (HttpRequest request, ILogger<Program> logger) =>
 {
-    // Настройки десериализации
-    var options = new JsonSerializerOptions
-    {
-        PropertyNameCaseInsensitive = true
-    };
-
-    // Чтение тела запроса и его парсинг
+    var constantKey = "WebAppData";
+    var botToken = builder.Configuration["BOT_TOKEN"]!;
+    var options = new JsonSerializerOptions();
+    
+    options.PropertyNameCaseInsensitive = true;
     var requestBody = await new StreamReader(request.Body).ReadToEndAsync();
     var payload = JsonSerializer.Deserialize<TgPayloadDto>(requestBody, options);
 
     if (payload == null || string.IsNullOrEmpty(payload.InitData))
     {
-        logger.LogError("Не удалось распарсить тело запроса.");
-        return Results.BadRequest("Неверные данные запроса.");
+        return Results.BadRequest();
     }
-
-    logger.LogInformation("Тело запроса, полученное от Telegram: {InitData}", payload.InitData);
-
-    // Парсинг строки init_data
     var data = HttpUtility.ParseQueryString(payload.InitData);
 
-    // Сортировка данных по алфавиту в SortedDictionary
-    var dataDict = new SortedDictionary<string, string>(
-        data.AllKeys.ToDictionary(x => x!, x => data[x]!),
-        StringComparer.Ordinal);
-
-    // Удаляем поле hash для формирования строки проверки данных
-    var dataCheckString = string.Join(
-        '\n', dataDict.Where(x => x.Key != "hash")
-        .Select(x => $"{x.Key}={x.Value}"));
-
-    logger.LogInformation("dataCheckString: {DataCheckString}", dataCheckString);
-
-    // Константный ключ для генерации секретного ключа
-    var constantKey = "WebAppData";
-
-    // Получаем токен бота из конфигурации
-    var botToken = builder.Configuration["BOT_TOKEN"];
-
-    // Генерация секретного ключа с использованием HMAC-SHA-256
-    var secretKey = HMACSHA256Hash(Encoding.UTF8.GetBytes(constantKey), Encoding.UTF8.GetBytes(botToken));
-
-    // Генерация хэша на основе строки проверки данных
-    var generatedHash = HMACSHA256Hash(secretKey, Encoding.UTF8.GetBytes(dataCheckString));
-
-    // Преобразование полученного хэша от Telegram в массив байтов
-    var actualHash = Convert.FromHexString(dataDict["hash"]);
-
-    // Сравнение вычисленного и полученного хэшей
-    if (actualHash.SequenceEqual(generatedHash))
-    {
-        logger.LogInformation("Данные подтверждены как подлинные.");
-        return Results.Ok(new { valid = true });
-    }
-    else
-    {
-        logger.LogError("Ошибка верификации.");
-        return Results.BadRequest("Ошибка верификации.");
-    }
+    return IsValidData(data, logger, constantKey, botToken) ? Results.Ok(new { valid = true }) : Results.BadRequest();
 });
 
 
@@ -261,15 +223,33 @@ bool IsValidUser(LoginModel login)
 }
 
 app.Run();
-byte[] HMACSHA256Hash(byte[] key, byte[] data)
+return;
+
+static byte[] Hmacsha256Hash(byte[] key, byte[] data)
 {
-    using (var hmac = new HMACSHA256(key))
-    {
-        return hmac.ComputeHash(data);
-    }
+    using var hmac = new HMACSHA256(key);
+    return hmac.ComputeHash(data);
 }
 
-// DTO для тела запроса
+static bool IsValidData(NameValueCollection nameValueCollection, ILogger<Program> logger1, string s, string botToken1)
+{
+    var dataDict = new SortedDictionary<string, string>(
+        nameValueCollection.AllKeys.ToDictionary(x => x!, x => nameValueCollection[x]!),
+        StringComparer.Ordinal);
+    var dataCheckString = string.Join(
+        '\n', dataDict.Where(x => x.Key != "hash")
+            .Select(x => $"{x.Key}={x.Value}"));
+
+    logger1.LogInformation("dataCheckString: {DataCheckString}", dataCheckString);
+
+    var secretKey = Hmacsha256Hash(Encoding.UTF8.GetBytes(s), Encoding.UTF8.GetBytes(botToken1));
+    var generatedHash = Hmacsha256Hash(secretKey, Encoding.UTF8.GetBytes(dataCheckString));
+    var actualHash = Convert.FromHexString(dataDict["hash"]);
+    
+    return actualHash.SequenceEqual(generatedHash);
+}
+
+
 public class TgPayloadDto
 {
     [JsonPropertyName("init_data")]
